refactor(global-search): move waitForIdle to utils
